updated tag and folder endpoints and seeded data

+++ routes/notes.js
@@ -4,14 +4,20 @@ const express = require('express');
 const router = express.Router();
 
 const mongoose = require('mongoose');
+const passport = require('passport');
 
 const Note = require('../models/note');
+// const Folder = require('../models/folder');
+// const Tag = require('../models/tag');
+
+router.use('/notes', passport.authenticate('jwt', {session:false, failWithError:true}));
 
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/notes', (req, res, next) => {
   const { searchTerm, folderId, tagId } = req.query;
+  const userId = req.user.id;
 
-  let filter = {};
+  let filter = { userId };
 
   /**
    * BONUS CHALLENGE - Search both title and content using $OR Operator
@@ -45,6 +51,8 @@ router.get('/notes', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/notes/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
+
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -52,7 +60,8 @@ router.get('/notes/:id', (req, res, next) => {
     return next(err);
   }
 
-  Note.findById(id)
+  Note.findById({_id: id, userId})
+    .select('title content created folderId tags')
     .populate('tags')
     .then(result => {
       if (result) {
@@ -69,13 +78,18 @@ router.get('/notes/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/notes', (req, res, next) => {
   const { title, content, folderId, tags } = req.body;
-
+  const userId = req.user.id;
+  
   /***** Never trust users - validate input *****/
   if (!title) {
     const err = new Error('Missing `title` in request body');
     err.status = 400;
     return next(err);
   }
+  
+  // if (!mongoose.Types.ObjectId.isValid(folderId)) {
+  //   createdNote.folderId=folderId;
+  // }
 
   if (tags) {
     tags.forEach((tag) => {
@@ -86,8 +100,7 @@ router.post('/notes', (req, res, next) => {
       }
     });
   }
-
-  const newItem = { title, content, folderId, tags };
+  const newItem = { title, content, folderId, tags, userId };
 
   Note.create(newItem)
     .then(result => {
@@ -102,7 +115,8 @@ router.post('/notes', (req, res, next) => {
 router.put('/notes/:id', (req, res, next) => {
   const { id } = req.params;
   const { title, content, folderId, tags } = req.body;
-
+  const userId = req.user.id;
+ 
   /***** Never trust users - validate input *****/
   if (!title) {
     const err = new Error('Missing `title` in request body');
@@ -129,9 +143,7 @@ router.put('/notes/:id', (req, res, next) => {
       }
     });
   }
-
-
-  const updateItem = { title, content, tags };
+  const updateItem = { title, content, tags, userId };
   const options = { new: true };
 
   Note.findByIdAndUpdate(id, updateItem, options)
@@ -151,10 +163,15 @@ router.put('/notes/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/notes/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
-  Note.findByIdAndRemove(id)
-    .then(() => {
-      res.status(204).end();
+  Note.findByIdAndRemove({ _id: id, userId})
+    .then(result => {
+      if(result){
+        res.status(204).end();
+      } else {
+        next();
+      }
     })
     .catch(err => {
       next(err);
@@ -162,3 +179,4 @@ router.delete('/notes/:id', (req, res, next) => {
 });
 
 module.exports = router;
+
